Refactor completely how user input is handled
Massive changes to how user input is handled. The input is now tokenized
and representative argc and argv is generated for it. Also, many
simplifications were made for the handling of arguments to the debugger
functions. Better error handling, while also having more static memory
allocations to avoid having to deal with heap as much as possible.
Important to clarify that no changes to the logic of the debugger have
been made. There are future changes.