Handle PIE executables
A lot of changes were made to accommodate the handling of PIE
executables. ASLR is disabled through the 'personality' syscall and
/proc/<tracee_pid>/maps is parsed to locate the start of the text
section once the program is actually loaded into memory. Meaning,
during ELF parsing, the type of the executable is now checked to
find whether it is PIE or not. Both cases can now be handled.